requirements.txt
aiohttp==3.9.5
aioredis==2.0.1
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anthropic==0.29.0
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
bcrypt==4.1.3
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.8
defusedxml==0.7.1
deprecation==2.1.0
distro==1.9.0
dnspython==2.6.1
email-validator==2.2.0
extra-streamlit-components==0.1.71
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.15.3
frozenlist==1.4.1
fsspec==2024.6.0
gitdb==4.0.11
gitpython==3.1.43
google-ai-generativelanguage==0.6.4
google-api-core==2.19.0
google-api-python-client==2.134.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-generativeai==0.5.4
googleapis-common-protos==1.63.1
gotrue==2.5.1
groq==0.9.0
grpcio==1.64.1
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
httpx-oauth==0.14.1
huggingface-hub==0.23.4
idna==3.7
importlib-metadata==8.0.0
jinja2==3.1.4
jiter==0.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
langchain==0.2.5
langchain-anthropic==0.1.15
langchain-core==0.2.9
langchain-google-genai==1.0.6
langchain-groq==0.1.5
langchain-openai==0.1.9
langchain-text-splitters==0.2.1
langsmith==0.1.81
litellm==1.41.13
markdown-it-py==3.0.0
markupsafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
ngrok==1.3.0
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
openai==1.35.3
orjson==3.10.5
packaging==24.1
pandas==2.2.2
pillow==10.3.0
postgrest==0.16.8
proto-plus==1.24.0
protobuf==4.25.3
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1-modules==0.4.0
pycparser==2.22
pydantic==2.7.4
pydantic-core==2.18.4
pydeck==0.9.1
pydrive2==1.19.0
pygments==2.18.0
pyjwt==2.8.0
pyngrok==7.1.6
pyopenssl==24.1.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
python-telegram-bot==21.3
pytz==2024.1
pyyaml==6.0.1
realtime==1.0.6
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sqlalchemy==2.0.31
starlette==0.37.2
storage3==0.7.6
streamlit==1.36.0
streamlit-authenticator==0.3.2
strenum==0.4.15
supabase==2.5.1
supafunc==0.4.6
tenacity==8.4.1
tiktoken==0.7.0
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
typer==0.12.3
typing-extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
yarl==1.9.4
zipp==3.19.2
